requirements.txt
pandas
numpy
matplotlib
argparse
biopython
more-itertools